Split windows firewall stuff into its own class
